+++ R/pivot_longer.R
@@ -13,6 +13,7 @@
 #' the same specification as `separate.()`.
 #' @param names_pattern If `names_to` contains multiple values, `names_pattern` takes
 #' the same specification as `extract.()`, a regular expression containing matching groups.
+#' @param names_prefix Append a prefix to the new names
 #' @param names_ptypes,values_ptypes A list of column name-prototype pairs. See ``?vctrs::`theory-faq-coercion```
 #' for more info on vctrs coercion.
 #' @param names_transform,values_transform A list of column name-function pairs. Use these arguments
@@ -42,8 +43,9 @@ pivot_longer. <- function(.df,
                           cols = everything(),
                           names_to = "name",
                           values_to = "value",
-                          names_sep = NULL,
                           names_pattern = NULL,
+                          names_sep = NULL,
+                          names_prefix = NULL,
                           names_ptypes = list(),
                           names_transform = list(),
                           names_repair = "check_unique",
@@ -60,8 +62,9 @@ pivot_longer..data.frame <- function(.df,
                                      cols = everything(),
                                      names_to = "name",
                                      values_to = "value",
-                                     names_sep = NULL,
                                      names_pattern = NULL,
+                                     names_sep = NULL,
+                                     names_prefix = NULL,
                                      names_ptypes = list(),
                                      names_transform = list(),
                                      names_repair = "check_unique",
@@ -111,6 +114,10 @@ pivot_longer..data.frame <- function(.df,
     variable.factor = fast_pivot,
     value.factor = FALSE
   )
+
+  if (!is.null(names_prefix)) {
+    .df[[var_name]] <- gsub(paste0("^", names_prefix), "", .df[[var_name]])
+  }
 
   if (multiple_names_to) {