+++ NEWS.md
@@ -3,6 +3,7 @@
 #### Functionality improvements
 * `mutate.()`: Variables are now updated in order when using ".by" (#166)
 * `summarize.()`: Variables are now created in order (#188)
+* `pivot_longer.()`: Added `names_prefix` arg
 
 # tidytable 0.5.8