+++ NEWS.md
@@ -9,6 +9,7 @@
 * `separate.()`: Added `convert` arg
 * `summarize.()`: Variables are now created in order (#188)
 * `pivot_longer.()`: Added `names_prefix` arg
+* `pivot_wider.()`: Can now use `".value"` in `names_glue`
 
 # tidytable 0.5.8