+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidytable 0.5.10 (in development)
+
 # tidytable 0.5.9
 
 #### New functions